Revert "fix: enable IAM auth for custom types (#2961) (#2971)"
This reverts commit 0e003d2656c46171c7fa2e947c92d701ede94f22.
