+++ packages/amplify-graphql-auth-transformer/src/__tests__/iam-custom-operations.test.ts
@@ -367,7 +367,8 @@ describe('Custom operations have @aws_iam directives when enableIamAuthorization
       expect(out.schema).not.toMatch(/onUpdateFooCustom: String.*@aws_iam/);
     });
 
-    test('Adds @aws_iam to non-model custom types when there is no model', () => {
+    // TODO: Enable this test once we fix https://github.com/aws-amplify/amplify-category-api/issues/2929
+    test.skip('Adds @aws_iam to non-model custom types when there is no model', () => {
       const strategy = makeStrategy(strategyType);
       const schema = /* GraphQL */ `
         type Foo {
@@ -402,7 +403,8 @@ describe('Custom operations have @aws_iam directives when enableIamAuthorization
       expect(out.schema).toMatch(/type Foo.*@aws_iam/);
     });
 
-    test('Adds @aws_iam to non-model custom types when there is a model', () => {
+    // TODO: Enable this test once we fix https://github.com/aws-amplify/amplify-category-api/issues/2929
+    test.skip('Adds @aws_iam to non-model custom types when there is a model', () => {
       const strategy = makeStrategy(strategyType);
       const schema = /* GraphQL */ `
         type Todo @model {
