+++ packages/amplify-graphql-transformer-core/API.md
@@ -363,9 +363,6 @@ export interface GraphQLTransformOptions {
 }
 
 // @public (undocumented)
-export const hasDirectiveWithName: (node: FieldDefinitionNode | InterfaceTypeDefinitionNode | ObjectTypeDefinitionNode, name: string) => boolean;
-
-// @public (undocumented)
 export type ImportAppSyncAPIInputs = {
     apiName: string;
     dataSourceConfig?: ImportedDataSourceConfig;
