+++ packages/amplify-graphql-transformer-core/src/index.ts
@@ -56,7 +56,6 @@ export {
   getSubscriptionFilterInputName,
   getTable,
   getType,
-  hasDirectiveWithName,
   isAmplifyDynamoDbModelDataSourceStrategy,
   isBuiltInGraphqlNode,
   isDefaultDynamoDbModelDataSourceStrategy,
